refine files hierarchy
