--- oneflow/user/kernels/broadcast_minimum_kernel_util.cpp
@@ -1,33 +1,0 @@
-/*
-Copyright 2020 The OneFlow Authors. All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-#include "oneflow/user/kernels/broadcast_minimum_kernel_util.h"
-
-namespace oneflow {
-namespace user_op {
-
-template<typename T>
-struct MinimumBackwardFunctor<DeviceType::kCPU, T> final {
-  void operator()(DeviceCtx* ctx, int64_t elem_cnt, const T* dz, const T* x, const T* y, T* dx,
-                  T* dy) {
-    DoUpdateMinimumGrad<T>(elem_cnt, dz, x, y, dx, dy);
-  }
-};
-
-template struct MinimumBackwardFunctor<DeviceType::kCPU, float>;
-template struct MinimumBackwardFunctor<DeviceType::kCPU, double>;
-
-}  // namespace user_op
-}  // namespace oneflow
